+++ src/main/java/io/github/srdjanv/tweakedpetroleumgas/api/mixins/IGasPumpjackAddons.java
@@ -1,9 +1,9 @@
-package io.github.srdjanv.tweakedpetroleumgas.api.util;
+package io.github.srdjanv.tweakedpetroleumgas.api.mixins;
 
 import mekanism.api.gas.Gas;
 import mekanism.api.gas.GasTank;
 import net.minecraft.util.EnumFacing;
-import io.github.srdjanv.tweakedpetroleum.api.ihelpers.IPumpjackAddons;
+import io.github.srdjanv.tweakedpetroleum.api.mixins.IPumpjackAddons;
 
 public interface IGasPumpjackAddons extends IPumpjackAddons {
