--- src/main/java/io/github/srdjanv/tweakedpetroleumgas/api/util/IGasReservoirType.java
@@ -1,10 +1,0 @@
-package io.github.srdjanv.tweakedpetroleumgas.api.util;
-
-import mekanism.api.gas.Gas;
-import io.github.srdjanv.tweakedpetroleum.api.ihelpers.IReservoirType;
-
-public interface IGasReservoirType extends IReservoirType {
-
-    Gas getGas();
-
-}
